niigata: move current status calculation into cell state
